pyproject.toml
[project]
name = "personal-website"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "flask>=3.1.0",
    "pytest>=8.3.5",
    "pytest-mock>=3.14.0",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
    "responses>=0.25.7",
]
